+++ mobile/test/java/io/envoyproxy/envoymobile/engine/testing/QuicTestServerTest.java
@@ -107,10 +107,10 @@ public class QuicTestServerTest {
 
   @Before
   public void setUpEngine() throws Exception {
-    QuicTestServer.startQuicTestServer();
+    TestJni.startQuicTestServer();
     CountDownLatch latch = new CountDownLatch(1);
-    engine = new AndroidEngineBuilder(
-                 appContext, new Custom(String.format(CONFIG, QuicTestServer.getServerPort())))
+    engine = new AndroidEngineBuilder(appContext,
+                                      new Custom(String.format(CONFIG, TestJni.getServerPort())))
                  .addLogLevel(LogLevel.WARN)
                  .setOnEngineRunning(() -> {
                    latch.countDown();
@@ -123,14 +123,15 @@ public class QuicTestServerTest {
   @After
   public void shutdownEngine() {
     engine.terminate();
-    QuicTestServer.shutdownQuicTestServer();
+    TestJni.shutdownQuicTestServer();
   }
 
   @Test
   public void get_simpleTxt() throws Exception {
     RequestScenario requestScenario = new RequestScenario()
                                           .setHttpMethod(RequestMethod.GET)
-                                          .setUrl(QuicTestServer.getServerURL() + "/simple.txt");
+                                          .addHeader("no_trailers", "true")
+                                          .setUrl(TestJni.getServerURL() + "/simple.txt");
 
     Response response = sendRequest(requestScenario);